refactor: 패키지를 분리합니다.
* DTO 객체를 request, response 패키지로 분리합니다.

* 인수 테스트의 steps와 추상화 한 공용 객체를 support 패키지로 이관합니다.
