+++ src/main/java/nextstep/subway/applicaion/PathService.java
@@ -1,6 +1,6 @@
 package nextstep.subway.applicaion;
 
-import nextstep.subway.applicaion.dto.PathResponse;
+import nextstep.subway.applicaion.dto.response.PathResponse;
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.Path;
 import nextstep.subway.domain.Station;
