+++ src/main/java/nextstep/subway/applicaion/dto/request/FavoriteRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.applicaion.dto.request;
 
 public class FavoriteRequest {
     private Long source;
