+++ src/main/java/nextstep/subway/applicaion/dto/request/LineRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.applicaion.dto.request;
 
 public class LineRequest {
     private String name;
