+++ src/test/java/nextstep/subway/acceptance/FavoriteAcceptanceTest.java
@@ -2,6 +2,8 @@ package nextstep.subway.acceptance;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import nextstep.subway.acceptance.support.AcceptanceTest;
+import nextstep.subway.acceptance.support.LineSteps;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -9,10 +11,15 @@ import org.junit.jupiter.api.Test;
 import java.util.HashMap;
 import java.util.Map;
 
-import static nextstep.subway.acceptance.FavoriteSteps.*;
-import static nextstep.subway.acceptance.LineSteps.지하철_노선에_지하철_구간_생성_요청;
-import static nextstep.subway.acceptance.MemberSteps.베어러_인증_로그인_요청;
-import static nextstep.subway.acceptance.StationSteps.지하철역_생성_요청;
+import static nextstep.subway.acceptance.support.FavoriteSteps.즐겨찾기_목록_조회_요청;
+import static nextstep.subway.acceptance.support.FavoriteSteps.즐겨찾기_목록_조회됨;
+import static nextstep.subway.acceptance.support.FavoriteSteps.즐겨찾기_삭제_요청;
+import static nextstep.subway.acceptance.support.FavoriteSteps.즐겨찾기_삭제됨;
+import static nextstep.subway.acceptance.support.FavoriteSteps.즐겨찾기_생성됨;
+import static nextstep.subway.acceptance.support.FavoriteSteps.즐겨찾기_생성을_요청;
+import static nextstep.subway.acceptance.support.LineSteps.지하철_노선에_지하철_구간_생성_요청;
+import static nextstep.subway.acceptance.support.MemberSteps.베어러_인증_로그인_요청;
+import static nextstep.subway.acceptance.support.StationSteps.지하철역_생성_요청;
 
 @DisplayName("즐겨찾기 관련 기능")
 public class FavoriteAcceptanceTest extends AcceptanceTest {
