+++ src/test/java/nextstep/subway/acceptance/support/LineSteps.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.acceptance.support;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
